pyproject.toml
[project]
name = "2bcloud"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.35",
    "gitpython>=3.1.46",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
